third_party/blink/web_tests/virtual/vertical-form-controls-direction-disabled-slider-vertical-enabled/css-parser/vertical-range-slider-appearance-expected.txt

CONSOLE WARNING: The keyword 'slider-vertical' specified to an 'appearance' property is not standardized. It will be removed in the future. Use <input type=range style="writing-mode: vertical-lr; direction: rtl"> instead.
